+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Next Up
+
+- Warrior of Elements Subclass parsing improvements.
+- Boon of Night Spirit and Tavern Brawler parsing improvements.
+
 # 6.0.21
 
 - Second Wind healing did not parse correctly.